Update Maven Module structure. Updated Colours
Changed Keschet into a Maven Multi-Module project. Extracted Colours out of Console and added GUI Hex colour values.
