+++ Base/src/main/kotlin/macro303/base/Direction.kt
@@ -1,0 +1,13 @@
+package macro303.base
+
+enum class Direction {
+	NORTH,
+	NORTH_EAST,
+	EAST,
+	SOUTH_EAST,
+	SOUTH,
+	SOUTH_WEST,
+	WEST,
+	NORTH_WEST,
+	INVALID
+}
