+++ pom.xml
@@ -5,9 +5,13 @@
 
 	<groupId>nz.co.macro303</groupId>
 	<artifactId>keschet</artifactId>
-	<version>2.1.1</version>
+	<version>2.1.2</version>
 	<name>Keschet</name>
-	<packaging>jar</packaging>
+	<packaging>pom</packaging>
+
+	<modules>
+		<module>Base</module>
+	</modules>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -33,11 +37,6 @@
 		<dependency>
 			<groupId>org.apache.logging.log4j</groupId>
 			<artifactId>log4j-core</artifactId>
-			<version>RELEASE</version>
-		</dependency>
-		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
 			<version>RELEASE</version>
 		</dependency>
 	</dependencies>
