--- src/main/kotlin/macro303/keschet/Direction.kt
@@ -1,13 +1,0 @@
-package macro303.keschet
-
-enum class Direction {
-	NORTH,
-	NORTH_EAST,
-	EAST,
-	SOUTH_EAST,
-	SOUTH,
-	SOUTH_WEST,
-	WEST,
-	NORTH_WEST,
-	INVALID
-}
