--- src/main/kotlin/macro303/keschet/Reader.kt
@@ -1,21 +1,0 @@
-package macro303.keschet
-
-import org.apache.logging.log4j.LogManager
-import java.util.*
-
-class Reader {
-	private val reader = Scanner(System.`in`)
-
-	fun readConsole(prompt: String): String {
-		Thread.sleep(10)
-		println(prompt)
-		print("> ")
-		val input = reader.nextLine().trim()
-		LOGGER.trace("String readConsole(String) = $input")
-		return input
-	}
-
-	companion object {
-		private val LOGGER = LogManager.getLogger(Reader::class.java)
-	}
-}
