--- src/main/kotlin/macro303/keschet/pieces/Archer.kt
@@ -1,22 +1,0 @@
-package macro303.keschet.pieces
-
-import macro303.keschet.Console
-import macro303.keschet.Direction
-
-class Archer(teamColour: Console.Colour) : Piece(teamColour = teamColour, maxDistance = 6, symbol = "A", validDirections = arrayOf(Direction.NORTH, Direction.EAST, Direction.SOUTH, Direction.WEST)) {
-
-	override fun equals(other: Any?): Boolean {
-		if (this === other) return true
-		if (other !is Archer) return false
-		if (!super.equals(other)) return false
-		return true
-	}
-
-	override fun hashCode(): Int {
-		return super.hashCode()
-	}
-
-	override fun toString(): String {
-		return "Archer() ${super.toString()}"
-	}
-}
