--- src/test/resources/log4j2.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<Configuration status = "WARN">
-	<Appenders>
-		<Console name = "Console" target="SYSTEM_OUT">
-			<PatternLayout pattern="%d{yyyy-MM-dd HH:mm:ss.SSS} %highlight{[%t] %-5level %logger{1} - %msg%n}{FATAL=white, ERROR=red, WARN=yellow, INFO=black, DEBUG=green, TRACE=blue}"/>
-		</Console>
-		<File name = "File" fileName="logs/keschet-test.log" append="false">
-			<PatternLayout pattern="%d{yyyy-MM-dd HH:mm:ss.SSS} [%t] %-5level %logger{1} - %msg%n"/>
-		</File>
-	</Appenders>
-	<Loggers>
-		<Root level="trace">
-			<AppenderRef ref="Console" level="warn" />
-			<AppenderRef ref="File" level="trace" />
-		</Root>
-	</Loggers>
-</Configuration>
